delete unecessary class
